+++ diabetes-dashboard/src/helpers/deleteMixin.js
@@ -1,4 +1,4 @@
-import Delete from "@/repositories/Delete.js";
+import Data from "@/repositories/Data.js";
 export const deleteMixin = {
     methods: {
         async deleteItem(parameters) {
@@ -6,7 +6,7 @@ export const deleteMixin = {
             if (this.$store.state.supervising.token) {
                 token = this.$store.state.supervising.token;
             }
-            Delete.post(parameters, token).then(
+            Data.deleteItem(parameters, token).then(
                 () => {
                     this.$toaster.showMessage({
                         message: "Delete is successful",
