--- diabetes-dashboard/src/repositories/DummyClient.js
@@ -1,9 +1,0 @@
-import axios from 'axios';
-
-const baseDomain = "https://gist.githubusercontent.com/";
-const baseURL = `${baseDomain}`;
-
-export default axios.create({
-    baseURL,
-    headers: {}
-});
